Move vsc-dependent tests to the right file.
